switched all uint256 to uint

+++ backend/blockchain/contracts/Market.sol
@@ -12,8 +12,8 @@ contract Market {
    ERC20 erc20;
    address _owner;
 
-   uint256 orderId;
-   uint256 productId;
+   uint orderId;
+   uint productId;
 
    /* Stakeholders */
    mapping(address => Procurer) procurers;
@@ -21,30 +21,30 @@ contract Market {
    mapping(address => Courier) couriers;
 
    /* Procurers' Purchase Orders */
-   mapping(uint256 => Structs.PurchaseOrder) orders;
+   mapping(uint => Structs.PurchaseOrder) orders;
 
    /* Suppliers' Products */
-   mapping(uint256 => Structs.Product) products;
+   mapping(uint => Structs.Product) products;
 
    /* ==================== Events ==================== */
    event Registered(string registerType, address registerer);
 
    /* === Supplier Events === */
-   event Listed(address supplier, address supplierEmployee, uint256 _productId, uint256 quantity, uint256 price);
-   event Unlisted(address supplier, address supplierEmployee, uint256 _productId);
-   event Relisted(address supplier, address supplierEmployee, uint256 _productId);
-   event ProductPriceUpdate(address supplier, address supplierEmployee, uint256 _productId, uint256 newPrice);
-   event ProductQuantityUpdate(address supplier, address supplierEmployee, uint256 _productId, uint256 newQuantity);
+   event Listed(address supplier, address supplierEmployee, uint _productId, uint quantity, uint price);
+   event Unlisted(address supplier, address supplierEmployee, uint _productId);
+   event Relisted(address supplier, address supplierEmployee, uint _productId);
+   event ProductPriceUpdate(address supplier, address supplierEmployee, uint _productId, uint newPrice);
+   event ProductQuantityUpdate(address supplier, address supplierEmployee, uint _productId, uint newQuantity);
 
    /* === Order Events === */
-   event OrderCreated(uint256 _orderId, address procurer, address logisticsEmployee, uint256 _productId, uint256 _quantity, uint256 price);
-   event OrderInternalApproved(uint256 _orderId, address procurer, address financeEmployee);
-   event OrderInternalRejected(uint256 _orderId, address procurer, address financeEmployee);
-   event OrderSupplierApproved(uint256 _orderId, address supplier, address supplierEmployee);
-   event OrderSupplierRejected(uint256 _orderId, address supplier, address supplierEmployee);
-   event OrderCourierAssigned(uint256 _orderId, address supplier, address supplierEmployee, address courier);
-   event OrderCourierDelivering(uint256 _orderId, address supplier, address courier, address courierEmployee);
-   event OrderDelivered(uint256 _orderId, address procurer);
+   event OrderCreated(uint _orderId, address procurer, address logisticsEmployee, uint _productId, uint _quantity, uint price);
+   event OrderInternalApproved(uint _orderId, address procurer, address financeEmployee);
+   event OrderInternalRejected(uint _orderId, address procurer, address financeEmployee);
+   event OrderSupplierApproved(uint _orderId, address supplier, address supplierEmployee);
+   event OrderSupplierRejected(uint _orderId, address supplier, address supplierEmployee);
+   event OrderCourierAssigned(uint _orderId, address supplier, address supplierEmployee, address courier);
+   event OrderCourierDelivering(uint _orderId, address supplier, address courier, address courierEmployee);
+   event OrderDelivered(uint _orderId, address procurer);
 
    /* ==================== Constructor ==================== */
 
@@ -83,7 +83,7 @@ contract Market {
     * @notice View a product listed in the marketplace.
     * @return Product
     */
-   function viewProduct(uint256 _productId) public view returns (Structs.Product memory) {
+   function viewProduct(uint _productId) public view returns (Structs.Product memory) {
       require(_productId > 0, "Invalid Product ID");
       require(products[_productId].productId != 0, "Product doesn't exist");
       require(products[_productId].listed, "Product is not listed currently");
@@ -96,7 +96,7 @@ contract Market {
     */
    function viewAllProducts() public view returns (Structs.Product[] memory) {
       Structs.Product[] memory _pa = new Structs.Product[](productId - 1);
-      for (uint256 i = 1; i < productId; i++) {
+      for (uint i = 1; i < productId; i++) {
          if (products[i].listed) {
             _pa[i - 1] = products[i];
          }
@@ -110,8 +110,8 @@ contract Market {
     */
    function viewSupplierProducts(address _supplier) public view returns (Structs.Product[] memory) {
       Structs.Product[] memory _pa = new Structs.Product[](productId - 1);
-      uint256 j = 0;
-      for (uint256 i = 1; i < productId; i++) {
+      uint j = 0;
+      for (uint i = 1; i < productId; i++) {
          if (products[i].listed && _supplier == products[i].supplier) {
             _pa[j++] = products[i];
          }
@@ -167,7 +167,7 @@ contract Market {
     * @notice View a purchase order. Can only be viewed by relevant stakeholders.
     * @return PurchaseOrder
     */
-   function viewPurchaseOrder(uint256 _orderId) public view returns (Structs.PurchaseOrder memory) {
+   function viewPurchaseOrder(uint _orderId) public view returns (Structs.PurchaseOrder memory) {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(
          orders[_orderId].procurer == msg.sender || orders[_orderId].supplier == msg.sender || orders[_orderId].courier == msg.sender,
@@ -183,7 +183,7 @@ contract Market {
     * @dev Called by a Procurer contract, from the logistics team only
     * @return Order ID of the newly created purchase order
     */
-   function createPurchaseOrder(uint256 _productId, uint256 quantity, uint256 price) public procurerOnly returns (uint256) {
+   function createPurchaseOrder(uint _productId, uint quantity, uint price) public procurerOnly returns (uint) {
       
       require(_productId > 0, "Invalid Product ID");
       require(quantity > 0, "Invalid Quantity");
@@ -221,7 +221,7 @@ contract Market {
     * @notice Procurer finance team approves a purchase order
     * @dev Called by a Procurer contract, from the finance team only
     */
-   function procurerApprovePurchaseOrder(uint256 _orderId) public procurerOnly {
+   function procurerApprovePurchaseOrder(uint _orderId) public procurerOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].procurer == msg.sender, "Only valid procurer can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.Ordered, "Current status of order is not ordered");
@@ -235,7 +235,7 @@ contract Market {
     * @notice Procurer finance team rejects a purchase order
     * @dev Called by a Procurer contract, from the finance team only
     */
-   function procurerRejectPurchaseOrder(uint256 _orderId) public procurerOnly {
+   function procurerRejectPurchaseOrder(uint _orderId) public procurerOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].procurer == msg.sender, "Only valid procurer can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.Ordered, "Current status of order is not ordered");
@@ -250,7 +250,7 @@ contract Market {
     * Will transfer the funds to the supplier.
     * @dev Called by a Procurer contract
     */
-   function deliveredByCourier(uint256 _orderId) public procurerOnly {
+   function deliveredByCourier(uint _orderId) public procurerOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].procurer == msg.sender, "Only valid procurer can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.Ordered, "Current status of order is not ordered");
@@ -269,8 +269,8 @@ contract Market {
     */
    function procurerViewAllPurchaseOrders() public view procurerOnly returns (Structs.PurchaseOrder[] memory) {
       Structs.PurchaseOrder[] memory _poa = new Structs.PurchaseOrder[](orderId - 1);
-      uint256 j = 0;
-      for (uint256 i = 1; i < orderId; i++) {
+      uint j = 0;
+      for (uint i = 1; i < orderId; i++) {
          if (orders[i].status != Structs.OrderStatus.notCreated && msg.sender == orders[i].procurer) {
             _poa[j++] = orders[i];
          }
@@ -284,7 +284,7 @@ contract Market {
     * @notice Supplier approves a purchase order, which waits for the supplier to assign a courier for delivery.
     * @dev Called by a Supplier contract
     */ 
-   function supplierApprovePurchaseOrder(uint256 _orderId) public supplierOnly {
+   function supplierApprovePurchaseOrder(uint _orderId) public supplierOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].supplier == msg.sender, "Only valid supplier can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.InternalApproved, "Current status of order is not internal approved");
@@ -301,7 +301,7 @@ contract Market {
     * @notice Supplier rejects a purchase order, hence the order will not have any further actions.
     * @dev Called by a Supplier contract
     */
-   function supplierRejectPurchaseorder(uint256 _orderId) public supplierOnly {
+   function supplierRejectPurchaseorder(uint _orderId) public supplierOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].supplier == msg.sender, "Only valid supplier can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.InternalApproved, "Current status of order is not internal approved");
@@ -320,7 +320,7 @@ contract Market {
     * @dev Called by a Supplier contract
     * @return Product ID
     */
-   function listProduct(uint256 quantityAvailable, uint256 price, string memory name) public supplierOnly returns (uint256) {      
+   function listProduct(uint quantityAvailable, uint price, string memory name) public supplierOnly returns (uint) {      
       Structs.Product memory _p = Structs.Product(
          msg.sender,
          productId,
@@ -344,7 +344,7 @@ contract Market {
     * Innately, the product is disabled and can be reenabled for further trade.
     * @dev Called by a Supplier contract
     */
-   function unlistProduct(uint256 _productId) public supplierOnly {
+   function unlistProduct(uint _productId) public supplierOnly {
       require(products[_productId].supplier != address(0), "Product does not exist");
       require(products[_productId].supplier == msg.sender, "Unauthorised supplier");
       products[_productId].listed = false;
@@ -356,7 +356,7 @@ contract Market {
     * @notice Supplier relists a product on the marketplace that has been previously unlisted.
     * @dev Called by a Supplier contract
     */
-   function relistProduct(uint256 _productId) public supplierOnly {
+   function relistProduct(uint _productId) public supplierOnly {
       require(products[_productId].supplier != address(0), "Product does not exist");
       require(products[_productId].supplier == msg.sender, "Unauthorised supplier");
       require(!products[_productId].listed, "Product listed, cannot be relisted");
@@ -369,7 +369,7 @@ contract Market {
     * @notice Supplier updates the price of a product on the marketplace
     * @dev Called by a Supplier contract
     */
-   function updateProductPrice(uint256 _productId, uint256 newPrice) public supplierOnly {
+   function updateProductPrice(uint _productId, uint newPrice) public supplierOnly {
       require(products[_productId].supplier != address(0), "Product does not exist");
       require(products[_productId].supplier == msg.sender, "Unauthorised supplier");
       products[_productId].price = newPrice;
@@ -381,7 +381,7 @@ contract Market {
     * @notice Supplier updates the price of a product on the marketplace
     * @dev Called by a Supplier contract
     */
-   function updateProductQuantity(uint256 _productId, uint256 newQuantity) public supplierOnly {
+   function updateProductQuantity(uint _productId, uint newQuantity) public supplierOnly {
       require(products[_productId].supplier != address(0), "Product does not exist");
       require(products[_productId].supplier == msg.sender, "Unauthorised supplier");
       products[_productId].quantityAvailable = newQuantity;
@@ -393,7 +393,7 @@ contract Market {
     * @notice View supplier's own product
     * @return Struct.Product
     */
-   function supplierViewSelfProduct(uint256 _productId) public view supplierOnly returns (Structs.Product memory) {
+   function supplierViewSelfProduct(uint _productId) public view supplierOnly returns (Structs.Product memory) {
       require(products[_productId].supplier != address(0), "Product does not exist");
       require(products[_productId].supplier == msg.sender, "Unauthorised supplier");
       return products[_productId];
@@ -405,8 +405,8 @@ contract Market {
     */
    function supplierViewAllSelfProducts() public view supplierOnly returns (Structs.Product[] memory) {
       Structs.Product[] memory _pa = new Structs.Product[](productId - 1);
-      uint256 j = 0;
-      for (uint256 i = 1; i < productId; i++) {
+      uint j = 0;
+      for (uint i = 1; i < productId; i++) {
          if (msg.sender == products[i].supplier) {
             _pa[j++] = products[i];
          }
@@ -418,7 +418,7 @@ contract Market {
     * @notice Supplier assigns a courier to deliver an order made by a procurer.
     * @dev Called by a Supplier contract
     */
-   function assignCourier(address courier, uint256 _orderId) public supplierOnly {
+   function assignCourier(address courier, uint _orderId) public supplierOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].supplier == msg.sender, "Only valid supplier can approve this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.SupplierApproved, "Current status of order is not supplier approved");
@@ -436,8 +436,8 @@ contract Market {
     */
    function supplierViewAllPurchaseOrders() public view supplierOnly returns (Structs.PurchaseOrder[] memory) {
       Structs.PurchaseOrder[] memory _poa = new Structs.PurchaseOrder[](orderId - 1);
-      uint256 j = 0;
-      for (uint256 i = 1; i < orderId; i++) {
+      uint j = 0;
+      for (uint i = 1; i < orderId; i++) {
          if (orders[i].status != Structs.OrderStatus.notCreated && msg.sender == orders[i].supplier) {
             _poa[j++] = orders[i];
          }
@@ -451,7 +451,7 @@ contract Market {
     * @notice Courier acknowledges they have received the goods to be delivered from the supplier.
     * @dev Called by a Courier contract
     */
-   function receivedByCourier(uint256 _orderId) public courierOnly {
+   function receivedByCourier(uint _orderId) public courierOnly {
       require(orders[_orderId].status != Structs.OrderStatus.notCreated, "Order does not exist");
       require(orders[_orderId].courier == msg.sender, "Only valid courier can access this purchase order");
       require(orders[_orderId].status == Structs.OrderStatus.SupplierApproved, "Current status of order is not supplier approved");
@@ -468,8 +468,8 @@ contract Market {
     */
    function courierViewAllPurchaseOrders() public view courierOnly returns (Structs.PurchaseOrder[] memory) {
       Structs.PurchaseOrder[] memory _poa = new Structs.PurchaseOrder[](orderId - 1);
-      uint256 j = 0;
-      for (uint256 i = 1; i < orderId; i++) {
+      uint j = 0;
+      for (uint i = 1; i < orderId; i++) {
          if (orders[i].status != Structs.OrderStatus.notCreated && msg.sender == orders[i].courier) {
             _poa[j++] = orders[i];
          }
